move to status-keycard-java 2.0.0
