+++ android/src/main/java/im/status/ethereum/keycard/Installer.java
@@ -3,31 +3,24 @@ package im.status.ethereum.keycard;
 import android.content.res.AssetManager;
 import android.util.Log;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.security.NoSuchAlgorithmException;
-import java.security.SecureRandom;
-import java.security.spec.InvalidKeySpecException;
+import im.status.keycard.globalplatform.GlobalPlatformCommandSet;
+import im.status.keycard.globalplatform.LoadCallback;
+import im.status.keycard.io.APDUException;
+import im.status.keycard.io.CardChannel;
+import org.bouncycastle.util.encoders.Hex;
 
-import im.status.hardwallet_lite_android.globalplatform.Load;
-import im.status.hardwallet_lite_android.io.APDUException;
-import im.status.hardwallet_lite_android.io.APDUResponse;
-import im.status.hardwallet_lite_android.io.CardChannel;
-import im.status.hardwallet_lite_android.globalplatform.ApplicationID;
-import im.status.hardwallet_lite_android.globalplatform.GlobalPlatformCommandSet;
+import java.io.IOException;
+import java.security.NoSuchAlgorithmException;
+import java.security.spec.InvalidKeySpecException;
 
 public class Installer {
     private CardChannel plainChannel;
     private AssetManager assets;
     private String capPath;
-    private static final String TAG = "SmartCardInstaller";
-
-    static final byte[] PACKAGE_AID = HexUtils.hexStringToByteArray("53746174757357616C6C6574");
-    static final byte[] WALLET_AID = HexUtils.hexStringToByteArray("53746174757357616C6C6574417070");
-    static final byte[] NDEF_APPLET_AID = HexUtils.hexStringToByteArray("53746174757357616C6C65744E4643");
-    static final byte[] NDEF_INSTANCE_AID = HexUtils.hexStringToByteArray("D2760000850101");
 
     private GlobalPlatformCommandSet cmdSet;
+
+    private static final String TAG = "SmartCardInstaller";
 
     public Installer(CardChannel channel, AssetManager assets, String capPath) {
         this.plainChannel = channel;
@@ -39,54 +32,30 @@ public class Installer {
         Log.i(TAG, "installation started...");
         long startTime = System.currentTimeMillis();
 
-        Log.i(TAG, "auto select sdaid...");
+        Log.i(TAG, "select ISD...");
         cmdSet = new GlobalPlatformCommandSet(this.plainChannel);
-        ApplicationID sdaid = new ApplicationID(cmdSet.select().checkOK().getData());
+        cmdSet.select().checkOK();
 
-        SecureRandom random = new SecureRandom();
-        byte hostChallenge[] = new byte[8];
-        random.nextBytes(hostChallenge);
-        Log.i(TAG, "initialize update...");
-        cmdSet.initializeUpdate(hostChallenge).checkOK();
+        Log.i(TAG, "opening secure channel...");
+        cmdSet.openSecureChannel();
 
-        Log.i(TAG, "external authenticate...");
-        cmdSet.externalAuthenticate(hostChallenge).checkOK();
+        Log.i(TAG, "deleting old version (if present)...");
+        cmdSet.deleteKeycardInstancesAndPackage();
 
-        Log.i(TAG, "delete NDEF instance AID...");
-        cmdSet.delete(NDEF_INSTANCE_AID).checkSW(APDUResponse.SW_OK, APDUResponse.SW_REFERENCED_DATA_NOT_FOUND);
+        Log.i(TAG, "loading package...");
+        cmdSet.loadKeycardPackage(this.assets.open(this.capPath), new LoadCallback() {
+            public void blockLoaded(int loadedBlock, int blockCount) {
+                Log.i(TAG, String.format("load %d/%d...", loadedBlock, blockCount));
+            }
+        });
 
-        Log.i(TAG, "delete wallet AID...");
-        cmdSet.delete(WALLET_AID).checkSW(APDUResponse.SW_OK, APDUResponse.SW_REFERENCED_DATA_NOT_FOUND);
+        Log.i(TAG, "installing NDEF applet...");
+        cmdSet.installNDEFApplet(Hex.decode("0024d40f12616e64726f69642e636f6d3a706b67696d2e7374617475732e657468657265756d")).checkOK();
 
-        Log.i(TAG, "delete package AID...");
-        cmdSet.delete(PACKAGE_AID).checkSW(APDUResponse.SW_OK, APDUResponse.SW_REFERENCED_DATA_NOT_FOUND);
-
-        Log.i(TAG, "install for load...");
-        cmdSet.installForLoad(PACKAGE_AID, sdaid.getAID()).checkSW(APDUResponse.SW_OK, APDUResponse.SW_REFERENCED_DATA_NOT_FOUND);
-
-        InputStream in = this.assets.open(this.capPath);
-        Load load = new Load(in);
-
-        java.util.Scanner s = new java.util.Scanner(in).useDelimiter("\\A");
-        Log.i(TAG, "cap file " + s.toString().length());
-
-        byte[] block;
-        int steps = load.blocksCount();
-        while((block = load.nextDataBlock()) != null) {
-            int count = load.getCount() - 1;
-            Log.i(TAG, String.format("load %d/%d...", count + 1, steps));
-            cmdSet.load(block, count, load.hasMore()).checkOK();
-        }
-
-        Log.i(TAG, "install for install ndef...");
-        byte[] params = HexUtils.hexStringToByteArray("0024d40f12616e64726f69642e636f6d3a706b67696d2e7374617475732e657468657265756d");
-        cmdSet.installForInstall(PACKAGE_AID, NDEF_APPLET_AID, NDEF_INSTANCE_AID, params).checkOK();
-
-        Log.i(TAG, "install for install wallet...");
-        cmdSet.installForInstall(PACKAGE_AID, WALLET_AID, WALLET_AID, new byte[0]).checkOK();
+        Log.i(TAG, "installing Keycard applet...");
+        cmdSet.installKeycardApplet().checkOK();
 
         long duration = System.currentTimeMillis() - startTime;
         Log.i(TAG, String.format("\n\ninstallation completed in %d seconds", duration / 1000));
     }
-
 }
